move random number generation into its own component
